put cache base classes in __init__
