+++ lib/ansible/plugins/cache/memcached.py
@@ -25,7 +25,7 @@ from itertools import chain
 
 from ansible import constants as C
 from ansible.errors import AnsibleError
-from ansible.plugins.cache.base import BaseCacheModule
+from ansible.plugins.cache import BaseCacheModule
 
 try:
     import memcache
